[refactor🔨] : front 작업중 변경사항 1차수정
