+++ src/main/java/shop/gaship/gashipshoppingmall/GashipShoppingMallApplication.java
@@ -3,6 +3,7 @@ package shop.gaship.gashipshoppingmall;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
+import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 /**
@@ -19,6 +20,7 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @SpringBootApplication
 @EnableScheduling
 @EnableAspectJAutoProxy
+@EnableAsync
 public class GashipShoppingMallApplication {
 
     public static void main(String[] args) {
