+++ src/main/java/shop/gaship/gashipshoppingmall/inquiry/controller/customer/CustomerInquiryRestController.java
@@ -1,5 +1,7 @@
 package shop.gaship.gashipshoppingmall.inquiry.controller.customer;
 
+import static shop.gaship.gashipshoppingmall.inquiry.inquiryenum.InquiryType.*;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -9,7 +11,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import shop.gaship.gashipshoppingmall.inquiry.dto.response.InquiryDetailsResponseDto;
 import shop.gaship.gashipshoppingmall.inquiry.dto.response.InquiryListResponseDto;
 import shop.gaship.gashipshoppingmall.inquiry.inquiryenum.InquiryType;
 import shop.gaship.gashipshoppingmall.inquiry.service.InquiryService;
@@ -17,7 +18,7 @@ import shop.gaship.gashipshoppingmall.util.PageResponse;
 import shop.gaship.gashipshoppingmall.statuscode.status.ProcessStatus;
 
 /**
- * 고객문의에 대한 독자적인 api를 처리하는 클래스입니다.
+ * 고객문의에 대한 요청을 처리하는 클래스입니다.
  * 현재는 조회기능만 있지만 차후 기능 확장시 해당 클래스에 고객문의관련 api 요청 처리에 대한 기능들이 추가됩니다.
  *
  * @author 최겸준
@@ -41,7 +42,7 @@ public class CustomerInquiryRestController {
     public ResponseEntity<PageResponse<InquiryListResponseDto>> customerInquiryList(
         @PageableDefault Pageable pageable) {
         Page<InquiryListResponseDto> inquiriesPage =
-            inquiryService.findInquiries(pageable, InquiryType.CUSTOMER_INQUIRIES.getValue());
+            inquiryService.findInquiries(pageable, CUSTOMER_INQUIRY.getValue());
 
         PageResponse<InquiryListResponseDto> pageResponse = new PageResponse<>(inquiriesPage);
 
@@ -60,7 +61,7 @@ public class CustomerInquiryRestController {
         Pageable pageable) {
         Page<InquiryListResponseDto> inquiriesPage =
             inquiryService.findInquiriesByStatusCodeNo(pageable,
-                InquiryType.CUSTOMER_INQUIRIES.getValue(), ProcessStatus.WAITING.getValue());
+                CUSTOMER_INQUIRY.getValue(), ProcessStatus.WAITING.getValue());
 
         PageResponse<InquiryListResponseDto> pageResponse = new PageResponse<>(inquiriesPage);
 
@@ -79,7 +80,7 @@ public class CustomerInquiryRestController {
         Pageable pageable) {
         Page<InquiryListResponseDto> inquiriesPage =
             inquiryService.findInquiriesByStatusCodeNo(pageable,
-                InquiryType.CUSTOMER_INQUIRIES.getValue(), ProcessStatus.COMPLETE.getValue());
+                CUSTOMER_INQUIRY.getValue(), ProcessStatus.COMPLETE.getValue());
 
         PageResponse<InquiryListResponseDto> pageResponse = new PageResponse<>(inquiriesPage);
 
@@ -99,25 +100,10 @@ public class CustomerInquiryRestController {
         Pageable pageable, @PathVariable Integer memberNo) {
         Page<InquiryListResponseDto> inquiriesPage =
             inquiryService.findInquiriesByMemberNo(pageable,
-                InquiryType.CUSTOMER_INQUIRIES.getValue(), memberNo);
+                CUSTOMER_INQUIRY.getValue(), memberNo);
 
         PageResponse<InquiryListResponseDto> pageResponse = new PageResponse<>(inquiriesPage);
 
         return ResponseEntity.ok(pageResponse);
-    }
-
-
-    /**
-     * 문의 상세조회 요청을 처리하는 기능입니다.
-     *
-     * @param inquiryNo 조회의 기준이 되는 문의번호입니다.
-     * @return ResponseEntity body에 InquiryDetailsResponseDto라는 상세조회에 필요한 정보가 담긴 객체를 넣어서 반환합니다.
-     * @author 최겸준
-     */
-    @GetMapping(value = "/{inquiryNo}")
-    public ResponseEntity<InquiryDetailsResponseDto> inquiryDetails(
-        @PathVariable Integer inquiryNo) {
-
-        return ResponseEntity.ok(inquiryService.findInquiry(inquiryNo));
     }
 }
