+++ src/test/java/shop/gaship/gashipshoppingmall/inquiry/controller/common/CommonInquiryControllerTest.java
@@ -40,6 +40,7 @@ import org.springframework.test.web.servlet.MvcResult;
 import shop.gaship.gashipshoppingmall.inquiry.controller.customer.CustomerInquiryRestController;
 import shop.gaship.gashipshoppingmall.inquiry.dto.request.InquiryAddRequestDto;
 import shop.gaship.gashipshoppingmall.inquiry.dto.request.InquiryAnswerRequestDto;
+import shop.gaship.gashipshoppingmall.inquiry.dto.response.InquiryDetailsResponseDto;
 import shop.gaship.gashipshoppingmall.inquiry.dto.response.InquiryListResponseDto;
 import shop.gaship.gashipshoppingmall.inquiry.service.InquiryService;
 import shop.gaship.gashipshoppingmall.statuscode.status.ProcessStatus;
@@ -539,6 +540,33 @@ class CommonInquiryControllerTest {
                 .content(objectMapper.writeValueAsString(inquiryAddRequestDtoWhenProduct))
                 .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isCreated());
+    }
+
+    @DisplayName("문의 단건조회요청을 잘 받고 service에 위임하여 반환된 InquiryDetailsResponseDto를 ResponseEntity의 body에 넣어서 반환한다. 상태코드 200")
+    @Test
+    void inquiryDetails() throws Exception {
+        InquiryDetailsResponseDto mockDto = new InquiryDetailsResponseDto(
+            1, 1, "memberNickName", "employeeName", "processStatus",
+            "productName", "title", "inquiryContent", LocalDateTime.now(), "answerContent",LocalDateTime.now(), null);
+
+        given(inquiryService.findInquiry(anyInt()))
+            .willReturn(mockDto);
+
+        MvcResult result = mvc.perform(get("/api/inquiries/{inquiryNo}", 1)
+                .accept(MediaType.APPLICATION_JSON))
+            .andExpect(status().isOk())
+            .andExpect(jsonPath("$.inquiryNo").value(mockDto.getInquiryNo()))
+            .andExpect(jsonPath("$.inquiryContent").value(mockDto.getInquiryContent()))
+            .andExpect(jsonPath("$.answerContent").value(mockDto.getAnswerContent()))
+            .andExpect(jsonPath("$.title").value(mockDto.getTitle()))
+            .andExpect(jsonPath("$.employeeName").value(mockDto.getEmployeeName()))
+            .andExpect(jsonPath("$.memberNickname").value(mockDto.getMemberNickname()))
+            .andExpect(jsonPath("$.productName").value(mockDto.getProductName()))
+            .andExpect(jsonPath("$.productNo").value(mockDto.getProductNo()))
+            .andExpect(jsonPath("$.processStatus").value(mockDto.getProcessStatus()))
+            .andReturn();
+
+        verify(inquiryService).findInquiry(1);
     }
 
 }
